clean up routes and test/dev code
