+++ app/components/appMainModule/appMainRoutes.js
@@ -46,35 +46,9 @@ function appMainRoutes($stateProvider) {
     }
   };
 
-  var localePage = {
-    name: 'appMain.localePage',
-    url: '^/:localeType/:localeId/:localeName',
-    // template: '<{{ctrl.localeType}}-page><' + '/' + '{{ctrl.localeType}}' + '-page>',
-    template: '<locale-page class="page-fade" locale-name="{{ctrl.localeName}}"></locale-page>',
-    controller: function($scope, $stateParams) {
-      var ctrl = this;
-      ctrl.localeName = $stateParams.localeName;
-      // ctrl.localeType = $stateParams.localeType;
-      // ctrl.localeType = 'city';
-    },
-    controllerAs: 'ctrl',
-    ncyBreadcrumb: {
-      // label: 'Locale Page',
-      label: '{{ctrl.localeName}}',
-      parent: 'appMain'
-    },
-    data: {
-      moduleClasses: 'page', // assign a module class to the <body> tag
-      pageClasses: 'localePage', // assign a page-specific class to the <body> tag
-      pageTitle: '{{ctrl.localeName}}', // set the title in the <head> section of the index.html file
-      pageDescription: 'Meta Description goes here' // meta description in <head>
-    }
-  };
-
   // $stateProvider.state(about);
   $stateProvider.state(city);
   $stateProvider.state(state);
-  $stateProvider.state(localePage);
 
 }
