+++ app/components/common/globalSearchBar/globalSearchBar.html
@@ -4,5 +4,5 @@
 
 <!--typeahead-popup-template-url="customPopupTemplate.html"-->
 
-<pre>Model: {{vm.selected | json}}</pre>
+<!--<pre>Model: {{vm.selected | json}}</pre>-->
